Set C binding for unix_getlogin_job in its own C file
