+++ src/unix/jbuild
@@ -194,6 +194,7 @@
     unix_readdir_n_job
     unix_readlink_job
     unix_lockf_job
+    unix_getlogin_job
     not_available
   ))
   (install_c_headers (
