+++ src/unix/lwt_unix_unix.h
@@ -43,50 +43,6 @@
 
 #include "unix_recv_send_utils.h"
 
-
-
-/* +-----------------------------------------------------------------+
-   | JOB: getlogin                                                   |
-   +-----------------------------------------------------------------+ */
-
-#if !defined(__ANDROID__)
-
-struct job_getlogin {
-    struct lwt_unix_job job;
-    char buffer[1024];
-    int result;
-};
-
-static void worker_getlogin(struct job_getlogin *job)
-{
-    job->result = getlogin_r(job->buffer, 1024);
-}
-
-static value result_getlogin(struct job_getlogin *job)
-{
-    int result = job->result;
-    if (result) {
-        lwt_unix_free_job(&job->job);
-        unix_error(result, "getlogin", Nothing);
-    } else {
-        value v = caml_copy_string(job->buffer);
-        lwt_unix_free_job(&job->job);
-        return v;
-    }
-}
-
-CAMLprim value lwt_unix_getlogin_job(value Unit)
-{
-    LWT_UNIX_INIT_JOB(job, getlogin, 0);
-    return lwt_unix_alloc_job(&job->job);
-}
-
-#else
-
-LWT_NOT_AVAILABLE1(unix_getlogin_job)
-
-#endif
-
 /* +-----------------------------------------------------------------+
    | JOBs: get{pw,gr}{nam,uid}                                       |
    +-----------------------------------------------------------------+ */
