+++ src/unix/lwt_unix_windows.h
@@ -131,7 +131,6 @@ LWT_NOT_AVAILABLE3(unix_getaddrinfo_job)
 LWT_NOT_AVAILABLE2(unix_getnameinfo_job)
 LWT_NOT_AVAILABLE1(unix_tcgetattr_job)
 LWT_NOT_AVAILABLE3(unix_tcsetattr_job)
-LWT_NOT_AVAILABLE1(unix_getlogin_job)
 LWT_NOT_AVAILABLE1(unix_getpwnam_job)
 LWT_NOT_AVAILABLE1(unix_getgrnam_job)
 LWT_NOT_AVAILABLE1(unix_getpwuid_job)
